Consolidate commonly used test helper functions
Move the logic into `testHelper.js` and individual
tests can import the functions accordingly
