+++ test/subSuites/httpTest.js
@@ -17,6 +17,7 @@ const chaiHttp = require('chai-http');
 chai.use(chaiHttp);
 const path = require('path');
 const fs = require('fs');
+const { fillRestRequest, addAndSend } = require('./testHelper');
 
 let electronApp,
   page,
@@ -48,84 +49,9 @@ module.exports = () => {
       }
     });
 
-    const fillRestRequest = async (
-      url,
-      method,
-      body = '',
-      headers = [],
-      cookies = []
-    ) => {
-      try {
-        // Make sure HTTP2 method is selected
-        await page.locator('button>> text=HTTP/2').click();
-
-        // click and select METHOD if it isn't GET
-        if (method !== 'GET') {
-          await page.locator('button#rest-method').click();
-          await page
-            .locator(`div[id^="composer"] >> a >> text=${method}`)
-            .click();
-        }
-
-        // type in url
-        await page.locator('#url-input').fill(url);
-
-        // set headers
-        headers.forEach(async ({ key, value }, index) => {
-          await page
-            .locator(`#header-row${index} >> [placeholder="Key"]`)
-            .fill(key);
-          await page
-            .locator(`#header-row${index} >> [placeholder="Value"]`)
-            .fill(value);
-          await page.locator('#add-header').click();
-        });
-
-        // set cookies
-        cookies.forEach(async ({ key, value }, index) => {
-          await page
-            .locator(`#cookie-row${index} >> [placeholder="Key"]`)
-            .fill(key);
-          await page
-            .locator(`#cookie-row${index} >> [placeholder="Value"]`)
-            .fill(value);
-          await page.locator('#add-cookie').click();
-        });
-
-        // Add BODY as JSON if it isn't GET
-        if (method !== 'GET') {
-          // select body type JSON
-          if ((await page.locator('#body-type-select').innerText()) === 'raw') {
-            await page.locator('#raw-body-type').click();
-            await page
-              .locator('.dropdown-item >> text=application/json')
-              .click();
-          }
-
-          // insert JSON content into body
-          const codeMirror = await page.locator('#body-entry-select');
-          await codeMirror.click();
-          const restBody = await codeMirror.locator('.cm-content');
-
-          try {
-            restBody.fill('');
-            await restBody.fill(body);
-          } catch (err) {
-            console.error(err);
-          }
-        }
-      } catch (err) {
-        console.error(err);
-      }
-    };
-
-    const addAndSend = async (num) => {
-      try {
-        await page.locator('button >> text=Add to Workspace').click();
-        await page.locator(`#send-button-${num}`).click();
-      } catch (err) {
-        console.error(err);
-      }
+    const fillAndSentRequest = async (url, method, n, body) => {
+      await fillRestRequest(page, url, method, body);
+      await addAndSend(page, n);
     };
 
     // The app takes a while to launch, and without these rendering checks
@@ -157,8 +83,7 @@ module.exports = () => {
           // TEST GET Request from JSON Placeholder
           const url = 'http://jsonplaceholder.typicode.com/posts';
           const method = 'GET';
-          await fillRestRequest(url, method);
-          await addAndSend(num++);
+          await fillAndSentRequest(url, method, num++);
           await new Promise((resolve) =>
             setTimeout(async () => {
               const statusCode = await page.locator('.status-tag').innerText();
@@ -200,8 +125,7 @@ module.exports = () => {
         try {
           const url = 'http://localhost:3004/book';
           const method = 'GET';
-          await fillRestRequest(url, method);
-          await addAndSend(num++);
+          await fillAndSentRequest(url, method, num++);
           await new Promise((resolve) =>
             setTimeout(async () => {
               const statusCode = await page.locator('.status-tag').innerText();
@@ -224,8 +148,7 @@ module.exports = () => {
           const method = 'POST';
           const body =
             '{"title": "HarryPotter", "author": "JK Rowling", "pages": 500}';
-          await fillRestRequest(url, method, body);
-          await addAndSend(num++);
+          await fillAndSentRequest(url, method, num++, body);
           await new Promise((resolve) =>
             setTimeout(async () => {
               const statusCode = await page.locator('.status-tag').innerText();
@@ -247,8 +170,7 @@ module.exports = () => {
           const url = 'http://localhost:3004/book/HarryPotter';
           const method = 'PUT';
           const body = '{"author": "Ron Weasley", "pages": 400}';
-          await fillRestRequest(url, method, body);
-          await addAndSend(num++);
+          await fillAndSentRequest(url, method, num++, body);
           await new Promise((resolve) =>
             setTimeout(async () => {
               const statusCode = await page.locator('.status-tag').innerText();
@@ -270,8 +192,7 @@ module.exports = () => {
           const url = 'http://localhost:3004/book/HarryPotter';
           const method = 'PATCH';
           const body = '{"author": "Hermoine Granger"}';
-          await fillRestRequest(url, method, body);
-          await addAndSend(num++);
+          await fillAndSentRequest(url, method, num++, body);
           await new Promise((resolve) =>
             setTimeout(async () => {
               const statusCode = await page.locator('.status-tag').innerText();
@@ -294,8 +215,7 @@ module.exports = () => {
           const url = 'http://localhost:3004/book/HarryPotter';
           const method = 'DELETE';
           const body = '{}';
-          await fillRestRequest(url, method, body);
-          await addAndSend(num++);
+          await fillAndSentRequest(url, method, num++, body);
           await new Promise((resolve) =>
             setTimeout(async () => {
               const statusCode = await page.locator('.status-tag').innerText();
@@ -314,8 +234,7 @@ module.exports = () => {
         try {
           const url = 'http://localhost:3004/book';
           const method = 'GET';
-          await fillRestRequest(url, method);
-          await addAndSend(num++);
+          await fillAndSentRequest(url, method, num++);
           await new Promise((resolve) =>
             setTimeout(async () => {
               const statusCode = await page.locator('.status-tag').innerText();
