+++ test/subSuites/testHelper.js
@@ -1,3 +1,159 @@
+const fillRestRequest = async (
+  page,
+  url,
+  method,
+  body = '',
+  headers = [],
+  cookies = []
+) => {
+  try {
+    // Make sure HTTP2 method is selected
+    await page.locator('button>> text=HTTP/2').click();
+
+    // click and select METHOD if it isn't GET
+    if (method !== 'GET') {
+      await page.locator('button#rest-method').click();
+      await page.locator(`div[id^="composer"] >> a >> text=${method}`).click();
+    }
+
+    // type in url
+    await page.locator('#url-input').fill(url);
+
+    // set headers
+    headers.forEach(async ({ key, value }, index) => {
+      await page
+        .locator(`#header-row${index} >> [placeholder="Key"]`)
+        .fill(key);
+      await page
+        .locator(`#header-row${index} >> [placeholder="Value"]`)
+        .fill(value);
+      await page.locator('#add-header').click();
+    });
+
+    // set cookies
+    cookies.forEach(async ({ key, value }, index) => {
+      await page
+        .locator(`#cookie-row${index} >> [placeholder="Key"]`)
+        .fill(key);
+      await page
+        .locator(`#cookie-row${index} >> [placeholder="Value"]`)
+        .fill(value);
+      await page.locator('#add-cookie').click();
+    });
+
+    // Add BODY as JSON if it isn't GET
+    if (method !== 'GET') {
+      // select body type JSON
+      if ((await page.locator('#body-type-select').innerText()) === 'raw') {
+        await page.locator('#raw-body-type').click();
+        await page.locator('.dropdown-item >> text=application/json').click();
+      }
+
+      // insert JSON content into body
+      const codeMirror = await page.locator('#body-entry-select');
+      await codeMirror.click();
+      const restBody = await codeMirror.locator('.cm-content');
+
+      try {
+        restBody.fill('');
+        await restBody.fill(body);
+      } catch (err) {
+        console.error(err);
+      }
+    }
+  } catch (err) {
+    console.error(err);
+  }
+};
+
+const fillGQLBasicInfo = async (
+  page,
+  url,
+  method,
+  headers = [],
+  cookies = []
+) => {
+  try {
+    // click and check GRAPHQL
+    await page.locator('button>> text=GraphQL').click();
+
+    // click and select METHOD if it isn't QUERY
+    if (method !== 'QUERY') {
+      await page.locator('button#graphql-method').click();
+      await page.locator(`div[id^="composer"] >> a >> text=${method}`).click();
+    }
+
+    // type in url
+    await page.locator('#url-input').fill(url);
+
+    // set headers
+    headers.forEach(async ({ key, value }, index) => {
+      await page
+        .locator(`#header-row${index} >> [placeholder="Key"]`)
+        .fill(key);
+      await page
+        .locator(`#header-row${index} >> [placeholder="Value"]`)
+        .fill(value);
+      await page.locator('#add-header').click();
+    });
+
+    // set cookies
+    cookies.forEach(async ({ key, value }, index) => {
+      await page
+        .locator(`#cookie-row${index} >> [placeholder="Key"]`)
+        .fill(key);
+      await page
+        .locator(`#cookie-row${index} >> [placeholder="Value"]`)
+        .fill(value);
+      await page.locator('#add-cookie').click();
+    });
+  } catch (err) {
+    console.error(err);
+  }
+};
+
+const fillGQLRequest = async (
+  page,
+  url,
+  method,
+  query = '',
+  variables = '',
+  headers = [],
+  cookies = []
+) => {
+  try {
+    await fillGQLBasicInfo(page, url, method, headers, cookies);
+
+    // select Body, clear it, and type in query
+    const codeMirror = await page.locator('#gql-body-entry');
+    await codeMirror.click();
+    const gqlBodyCode = await codeMirror.locator('.cm-content');
+
+    try {
+      await gqlBodyCode.fill('');
+      await gqlBodyCode.fill(query);
+    } catch (err) {
+      console.error(err);
+    }
+
+    // select Variables and type in variables
+    const codeMirror2 = await page.locator('#gql-var-entry');
+    await codeMirror2.click();
+    await codeMirror2.locator('.cm-content').fill(variables);
+  } catch (err) {
+    console.error(err);
+  }
+};
+
+const addAndSend = async (page, n) => {
+  try {
+    await page.locator('button >> text=Add to Workspace').click();
+    await page.locator(`#send-button-${n}`).click();
+  } catch (err) {
+    console.error(err);
+  }
+};
+
 // Bring in the Clear & Fill Test Script Area for improved code readability.
 const clearAndFillTestScriptArea = async (page, script) => {
   try {
@@ -23,5 +179,11 @@ const clearAndFillTestScriptArea = async (page, script) => {
   }
 };
 
-module.exports = { clearAndFillTestScriptArea };
+module.exports = {
+  fillRestRequest,
+  fillGQLBasicInfo,
+  fillGQLRequest,
+  addAndSend,
+  clearAndFillTestScriptArea,
+};
